Version 4: soporte para coeficientes decimales usando double

+++ src/Main.java
@@ -4,19 +4,17 @@ import java.util.Scanner;
 
 public class Main {
     public static void main(String[] args) {
-
-        // Version 3: validacion de discriminante y control de ecuacion invalida
+        // Version 4: Se mejoran los tipos de dato para aceptar coeficientes decimales.
         try (Scanner sc = new Scanner(System.in)) {
-
-            int a, b, c;
+            double a, b, c;
             double discriminante, x1, x2;
 
             System.out.print("Ingrese a: ");
-            a = sc.nextInt();
+            a = sc.nextDouble();
             System.out.print("Ingrese b: ");
-            b = sc.nextInt();
+            b = sc.nextDouble();
             System.out.print("Ingrese c: ");
-            c = sc.nextInt();
+            c = sc.nextDouble();
 
             if (a == 0) {
                 System.out.println("No es una ecuacion de segundo grado porque a no puede ser 0.");
